Revert "move getECCEnv to events-api"
This reverts commit 097104b811b7690ab06a098e95580438a87fad59.
