+++ ecc/blocks/form-handler/controllers/venue-info-component-controller.js
@@ -1,8 +1,7 @@
 /* eslint-disable no-unused-vars */
 import { createVenue, replaceVenue } from '../../../scripts/esp-controller.js';
 import BlockMediator from '../../../scripts/deps/block-mediator.min.js';
-import { changeInputValue, getSecret } from '../../../scripts/utils.js';
-import { getECCEnv } from '../../../scripts/helpers.js';
+import { changeInputValue, getECCEnv, getSecret } from '../../../scripts/utils.js';
 import { buildErrorMessage } from '../form-handler.js';
 
 function togglePrefillableFieldsHiddenState(component, showPrefilledFields) {
