+++ ecc/scripts/esp-controller.js
@@ -1,4 +1,4 @@
-import { getECCEnv } from './helpers.js';
+import { getECCEnv } from './utils.js';
 
 export const getCaasTags = (() => {
   let cache;
